+++ framework/src/operators/foreach_batch.rs
