+++ framework/src/operators/groupby_batch.rs
